+++ src/components/OverlayManager/Login/queries.ts
@@ -7,12 +7,13 @@ import {
 } from "./types/RegisterAccount";
 
 const accountRegisterMutation = gql`
-  mutation RegisterAccount($email: String!, $password: String!) {
-    accountRegister(input: { email: $email, password: $password }) {
+  mutation RegisterAccount($email: String!, $password: String!, $redirectUrl: String!) {
+    accountRegister(input: { email: $email, password: $password, redirectUrl: $redirectUrl }) {
       errors {
         field
         message
       }
+      requiresConfirmation
     }
   }
 `;