+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable, unreleased changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Account confirmation mechanism - #565 by @tomaszszymanski129
 - Add missing product attributes on product page - #536 by @orzechdev
 - Change register mutation to accountRegister - #549 by @tomaszszymanski129
 - Add `ProductVariantPicker` component supporting multiple product variant attributes - #550 by @orzechdev